+++ src/filterBar/index.tsx
@@ -1,0 +1,2 @@
+export { FilterBar, FilterBarProps } from './FilterBar'
+export { FilterBar as default } from './FilterBar'