+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - **[UPDATE]** Add `noHorizontalSpacing` prop to disable horizontal padding on `Modal` component.
 [...]
+- **[NEW]** Added `FilterBar` component
 
 # v37.1.1 (22/07/2020)